requirements.txt
# ============================================================
# BTTS Slip AI Dashboard — Python Dependencies
# ============================================================
# Install all with:
#   pip install -r requirements.txt

# --- Core UI ---
streamlit>=1.32.0

# --- Data & Math ---
pandas>=2.0.0
numpy>=1.24.0

# --- ML (Phase 1: scikit-learn heuristics) ---
scikit-learn>=1.3.0

# --- ML Upgrade (Phase 2: XGBoost) ---
xgboost>=2.0.0

# --- API integration (Phase 2) ---
requests>=2.31.0

# --- Utilities ---
joblib>=1.3.0       # model serialisation (used in Phase 2)
python-dotenv>=1.0.0  # .env file for API key management
